Improve install script to work out of the box

+++ win10/install.ps1
@@ -19,9 +19,7 @@ $env:Dotfiles_PowershellHome = $PowershellHomeLocation
 $env:Dotfiles_MachineScripts = $MachineScriptsLocation
 [Environment]::SetEnvironmentVariable("Dotfiles_MachineScripts", $env:Dotfiles_MachineScripts, "User")
 
-
 # Install and update dependencies
-cup chocolatey
 cup all
 choco install `
 	dotnet4.7.1 `
@@ -35,13 +33,22 @@ choco install `
 	putty.install `
 	visualstudiocode `
 	conemu `
-    -y
+	7zip.install `
+	-y
+
+$env:Dotfiles_MachineScripts = $MachineScriptsLocation
+[Environment]::SetEnvironmentVariable("Dotfiles_MachineScripts", $env:Dotfiles_MachineScripts, "User")
 
 # Link configs
 Function SymlinkConfig($DotfilesRelativePath, $DestinationPath) {
 	$DotfilesAbsolutePath = Join-Path $env:Dotfiles $DotfilesRelativePath
-	Remove-Item -Path $DestinationPath
+	If (Test-Path $DestinationPath) {
+		Remove-Item -Path $DestinationPath
+	}
 	New-Item -Path $DestinationPath -ItemType SymbolicLink -Value $DotfilesAbsolutePath
 }
 SymlinkConfig "win10\conemu\ConEmu.xml" "$env:AppData\ConEmu.xml"
 SymlinkConfig "git\.gitconfig" "$env:HOMEDRIVE$env:HOMEPATH\.gitconfig"
+
+# Install Powershell profile
+Write-Output "Invoke-Expression `". `$env:Dotfiles\win10\powershell\profile.ps1`"" > $PROFILE
